add edit and remove to sale order

+++ pos/imports/ui/pages/order.js
@@ -36,14 +36,15 @@ import './info-tab.html';
 //methods
 import {saleOrderInfo} from '../../../common/methods/sale-order.js'
 import {customerInfo} from '../../../common/methods/customer.js';
+import {isInvoiceExist} from '../../../common/methods/sale-order';
 //Tracker for customer infomation
-Tracker.autorun(function(){
-  if(Session.get("saleOrderCustomerId")){
-    customerInfo.callPromise({_id: Session.get("saleOrderCustomerId")})
-    .then(function(result){
-      Session.set('customerInfo', result);
-    })
-  }
+Tracker.autorun(function () {
+    if (Session.get("saleOrderCustomerId")) {
+        customerInfo.callPromise({_id: Session.get("saleOrderCustomerId")})
+            .then(function (result) {
+                Session.set('customerInfo', result);
+            })
+    }
 });
 // Declare template
 let indexTmpl = Template.Pos_order,
@@ -76,15 +77,27 @@ indexTmpl.events({
         alertify.order(fa('plus', TAPi18n.__('pos.order.title')), renderTemplate(newTmpl)).maximize();
     },
     'click .js-update' (event, instance) {
-        alertify.order(fa('pencil', TAPi18n.__('pos.order.title')), renderTemplate(editTmpl, this));
+        Meteor.call("pos.isInvoiceExist", {_id: this._id}, (err, result)=> {
+            if(result.exist){
+                swal('បញ្ជាក់!', `សូមធ្វើការលុប Invoice លេខ​ ${result.invoiceId} ជាមុនសិន!​​​​`, 'error');
+            }else{
+                alertify.order(fa('pencil', TAPi18n.__('pos.order.title')), renderTemplate(editTmpl, this));
+            }
+        });
     },
     'click .js-destroy' (event, instance) {
         let data = this;
-        destroyAction(
-            Order,
-            {_id: data._id},
-            {title: TAPi18n.__('pos.order.title'), itemTitle: data._id}
-        );
+        Meteor.call("pos.isInvoiceExist", {_id: this._id}, (err, result)=> {
+            if(result.exist){
+                swal('បញ្ជាក់!', `សូមធ្វើការលុប Invoice លេខ​ ${result.invoiceId} ជាមុនសិន!​​​​`, 'error');
+            }else{
+                destroyAction(
+                    Order,
+                    {_id: data._id},
+                    {title: TAPi18n.__('pos.order.title'), itemTitle: data._id}
+                );
+            }
+        });
     },
     'click .js-display' (event, instance) {
         alertify.orderShow(fa('eye', TAPi18n.__('pos.order.title')), renderTemplate(showTmpl, this));
@@ -100,40 +113,40 @@ indexTmpl.events({
 
 // New
 newTmpl.events({
-  'change [name=customerId]'(event, instance){
-    if(event.currentTarget.value != ''){
-      Session.set('saleOrderCustomerId', event.currentTarget.value);
+    'change [name=customerId]'(event, instance){
+        if (event.currentTarget.value != '') {
+            Session.set('saleOrderCustomerId', event.currentTarget.value);
+        }
     }
-  }
 })
 newTmpl.helpers({
-  customerInfo() {
-    let customerInfo = Session.get('customerInfo');
-    if(!customerInfo){
-      return {empty: true, message: 'No data available'}
-    }
+    customerInfo() {
+        let customerInfo = Session.get('customerInfo');
+        if (!customerInfo) {
+            return {empty: true, message: 'No data available'}
+        }
 
-    return {
-      fields: `<li>Phone: <b>${customerInfo.telephone ? customerInfo.telephone : ''}</b></li>
+        return {
+            fields: `<li>Phone: <b>${customerInfo.telephone ? customerInfo.telephone : ''}</b></li>
               <li>Opening Balance: <span class="label label-success">0</span></li>
               <li >Credit Limit: <span class="label label-warning">${customerInfo.creditLimit ? numeral(customerInfo.creditLimit).format('0,0.00') : 0}</span></li>
               <li>Sale Order to be invoice: <span class="label label-primary">0</span>`
-  };
-  },
-  collection(){
-      return Order;
-  },
-  itemsCollection(){
-      return itemsCollection;
-  },
-  disabledSubmitBtn: function () {
-      let cont = itemsCollection.find().count();
-      if (cont == 0) {
-          return {disabled: true};
-      }
+        };
+    },
+    collection(){
+        return Order;
+    },
+    itemsCollection(){
+        return itemsCollection;
+    },
+    disabledSubmitBtn: function () {
+        let cont = itemsCollection.find().count();
+        if (cont == 0) {
+            return {disabled: true};
+        }
 
-      return {};
-  }
+        return {};
+    }
 });
 
 newTmpl.onDestroyed(function () {
@@ -159,10 +172,10 @@ editTmpl.helpers({
 
         // Add items to local collection
         _.forEach(data.items, (value)=> {
-          Meteor.call('getItem', value.itemId, function(err, result){
-            value.name = result.name;
-            itemsCollection.insert(value);
-          })
+            Meteor.call('getItem', value.itemId, function (err, result) {
+                value.name = result.name;
+                itemsCollection.insert(value);
+            })
         });
 
         return data;
@@ -187,16 +200,16 @@ editTmpl.onDestroyed(function () {
 
 // Show
 showTmpl.onCreated(function () {
-  this.saleOrder = new ReactiveVar();
-  this.autorun(()=> {
-      saleOrderInfo.callPromise({_id: this.data._id})
-            .then( (result) => {
-              this.saleOrder.set(result);
+    this.saleOrder = new ReactiveVar();
+    this.autorun(()=> {
+        saleOrderInfo.callPromise({_id: this.data._id})
+            .then((result) => {
+                this.saleOrder.set(result);
             }).catch(function (err) {
                 console.log(err.message);
             }
         );
-  });
+    });
 });
 
 showTmpl.helpers({
